Moved CourseStudyModeController
