+++ app/views/candidate_interface/course_choices/course_study_mode/_form_fields.html.erb
