+++ app/views/candidate_interface/course_choices/course_study_mode/new.html.erb
